This is PCB_V05 soft that debug the second stop , Data:2019-04-09,Version: S1

+++ nHIOS_PIC16F1828_Sample_pcv05.X/Delay.h
@@ -13,7 +13,7 @@ extern "C" {
 #endif
  void delay_100us(uint t);
  void delay_1ms(uint t);
- void delay_10ms(ulong t);
+ void delay_10ms(uint t);
  void init_fosc(void);
